# Minihanked-1.0.0-swagger.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: 'http://localhost:8080'
    description: Local development
  - url: 'https://minihanked.variksoo.ee/'
    description: CI/CD Live /w domain and HTTPS
...
paths:
  '/api/procurements/{id}':
    get:
      tags:
        - miniprocurement-controller
      operationId: getById_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MiniProcurementDto'
    put:
      tags:
        - miniprocurement-controller
      operationId: updateProcurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiniProcurementDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MiniProcurementDto'
    delete:
      tags:
        - miniprocurement-controller
      operationId: deleteProcurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
      deprecated: true
  '/api/miniprocurementPartners/{id}':
    summary: Tie together Procurements and Partners in a 0 or 1 procurements to 0 to many partners.
    description: This was created by Kristjan Variksoo.
    get:
      "summary": "Find miniprocurement and partner links by ID"
      tags:
        - miniprocurement-partner-controller
      operationId: getById_2
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MiniprocurementPartnerDtoDto'
    put:
      tags:
        - miniprocurement-partner-controller
      operationId: updateProcurement
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MiniprocurementPartnerDtoDto'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/MiniprocurementPartnerDtoDto'
    delete:
      tags:
        - miniprocurement-partner-controller
      operationId: deleteProcurementPartner
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
      deprecated: true
...
  '/api/partner/{id}':
    get:
      tags:
        - partner-controller
      operationId: getById_4
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PartnerDto'
...
components:
  schemas:
    MiniProcurementDto:
      required:
        - addedBy
        - deadline
        - procurementName
      type: object
      properties:
        procurementId:
          type: integer
          format: int32
        procurementName:
          type: string
        amount:
          type: integer
          format: int32
        description:
          type: string
        requirements:
          type: string
        contractId:
          type: integer
          format: int32
        timeAdded:
          type: string
          format: date-time
        addedBy:
          type: integer
          format: int32
        deadline:
          type: string
          format: date-time
        status:
          type: integer
          format: int32
        timeFinished:
          type: string
          format: date-time
...
    PartnerDto:
      type: object
      properties:
        personID:
          type: integer
          format: int32
        personName:
          type: string
        timeOfRegister:
          type: string
          format: date-time
        regNr:
          type: integer
          format: int64
        partnerInfo:
          type: string
        email:
          type: string
...